requirements.txt
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
pytz==2024.1
python-dateutil==2.9.0.post0